Add files via upload
These are Yaml definitions to create a pod with what type of instance and defining replication definition using  v1 ReplicationController and apps/v1 ReplicaSet. This is to understand how to nest the yaml file and understanding the basics of root level fields (root-keys) in a yaml for Kubernetes, Terraform, Ansible, or CloudFormation.