Add files via upload
Learned how webhooks works with K8s, how webhook deployment is made and how admission controller works in order to admit the creation of a pod.
How and what is mutating and validating admission controller works. Adding secrets to K8s cluster and creating service for the webhook deployment. so, that it is exposed on the port.